backend/requirements.txt
Flask==3.1.1
mysql-connector-python==9.3.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
bcrypt==4.0.1
flask-cors==4.0.0